add interface for project export configuration
